requirements.txt
# DeepFake Detection System Requirements
# Layer 0-9 Implementation Dependencies

# Core ML
numpy>=1.21.0
scipy>=1.7.0

# Image/Video Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Audio Processing
librosa>=0.9.0
soundfile>=0.10.0

# Deep Learning (Optional - for neural detector)
torch>=1.9.0
torchvision>=0.10.0

# Cryptography (Layer 5 - Provenance)
cryptography>=3.4.0

# API (Layer 6-7)
flask>=2.0.0  # Or fastapi>=0.68.0
requests>=2.26.0

# Data Handling
pandas>=1.3.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Type Checking
mypy>=0.910

# Code Quality
black>=21.7b0
flake8>=3.9.0
